fix(player): stop the home skeleton hero sitting below the app bar
The loading list carried a top padding of safeAreaTop + kToolbarHeight
while the loaded CustomScrollView carried none under a Scaffold with
extendBodyBehindAppBar, so the hero jumped up by roughly 80 to 100px
when data landed. It also spaced its rails apart where the loaded
layout does not.

The loading branch moves to HomeLoadingSkeleton so it can be pumped in
a test without faking a Riverpod stream.
